implement home page

+++ Domains/ApplicationDomain/Identity/Entities/User.cs
@@ -10,7 +10,7 @@ namespace ApplicationDomain.Identity.Entities
     public class User : IdentityUser<int>
     {
 		public string Fullname { set; get; }
-		public float Height { set; get; }
+		public int Height { set; get; }
 		public int YearOfBirth { set; get; }
 		public Gender Gender { set; get; }
 		public DateTime DateJoined { set; get; }
